chore: update controllers and domain to use sign out instead of logout
