+++ apps/rest_api/lib/controllers/public/auth.ex
@@ -1,4 +1,4 @@
-defmodule RestAPI.Controllers.Public.Tokens do
+defmodule RestAPI.Controllers.Public.Auth do
   @moduledoc false
 
   use RestAPI.Controller, :controller
@@ -43,6 +43,30 @@ defmodule RestAPI.Controllers.Public.Tokens do
 
       {:error, _reason} = error ->
         error
+    end
+  end
+
+  @doc "Logout the authenticated subject session."
+  @spec sign_out(conn :: Plug.Conn.t(), params :: map()) :: Plug.Conn.t()
+  def sign_out(%{private: %{session: session}} = conn, _params) do
+    session.jti
+    |> Commands.sign_out_session()
+    |> case do
+      {:ok, _count} -> send_resp(conn, :no_content, "")
+      {:error, :not_active} -> send_resp(conn, :forbidden, "")
+      {:error, :not_found} -> send_resp(conn, :not_found, "")
+    end
+  end
+
+  @doc "Logout subject authenticated sessions."
+  @spec sign_out_all_sessions(conn :: Plug.Conn.t(), params :: map()) :: Plug.Conn.t()
+  def sign_out_all_sessions(%{private: %{session: session}} = conn, _params) do
+    session.subject_id
+    |> Commands.sign_out_all_sessions(session.subject_type)
+    |> case do
+      {:ok, _count} -> send_resp(conn, :no_content, "")
+      {:error, :not_active} -> send_resp(conn, :forbidden, "")
+      {:error, :not_found} -> send_resp(conn, :not_found, "")
     end
   end
 end
